% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcrle}
\alias{calcrle}
\title{RLE function}
\usage{
calcrle(expTable, logTable = FALSE, plotTable = FALSE)
}
\arguments{
\item{expTable}{Expression table.}

\item{logTable}{If expression table is already on log scale. Defaults to FALSE.}

\item{plotTable}{Plot results. Defaults to FALSE.}
}
\description{
This function calculates RLE and plots.
}
\examples{
calcrle()
}
\keyword{RLE}
