% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clib.R
\name{clib_bin_install}
\alias{clib_bin_install}
\title{clib_bin_install
Install the clib binaries to a specified directory
this is usefull to get clib into your PATH}
\usage{
clib_bin_install(dest_dir)
}
\arguments{
\item{dest_dir}{Destination directory to install the binaries}
}
\description{
clib_bin_install
Install the clib binaries to a specified directory
this is usefull to get clib into your PATH
}
