% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clib.R
\name{strip_ansi}
\alias{strip_ansi}
\title{Strip ANSI escape codes from a character vector}
\usage{
strip_ansi(x)
}
\arguments{
\item{x}{Character vector possibly containing ANSI escape codes}
}
\value{
Character vector with ANSI codes removed
}
\description{
Strip ANSI escape codes from a character vector
}
