% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clib.R
\name{clib_run}
\alias{clib_run}
\title{clib run
Run a clib command}
\usage{
clib_run(command, args)
}
\arguments{
\item{command}{Name of the clib command to run}

\item{args}{Arguments to pass to clib}
}
\value{
Output of the clib command
}
\description{
clib run
Run a clib command
}
